+++ .github/workflows/asw2122.yml
@@ -46,7 +46,7 @@ jobs:
   docker-push-webapp:
     name: Push webapp Docker Image to GitHub Packages
     runs-on: ubuntu-latest
-    needs: [unit-test-restapi]
+    needs: [unit-test-webapp, unit-test-restapi]
     steps:
       - uses: actions/checkout@v2
       - name: Publish to Registry
@@ -64,7 +64,7 @@ jobs:
   docker-push-restapi:
     name: Push restapi Docker Image to GitHub Packages
     runs-on: ubuntu-latest
-    needs: [unit-test-restapi]
+    needs: [unit-test-restapi, unit-test-webapp]
     steps:
       - uses: actions/checkout@v2
       - name: Publish to Registry
