+++ webapp/package.json
@@ -50,8 +50,7 @@
   },
   "jest": {
     "collectCoverageFrom": [
-      "./src/components/*.tsx",
-      "./e2e/steps/*.ts"
+      "./src/components/*.tsx"
     ]
   },
   "eslintConfig": {
